+++ htfs/commands.go
@@ -23,6 +23,14 @@ type CatalogPuller func(string, string, bool) error
 func NewEnvironment(condafile, holozip string, restore, force bool, puller CatalogPuller) (label string, scorecard common.Scorecard, err error) {
 	defer fail.Around(&err)
 
+	if common.WarrantyVoided() {
+		tree, err := New()
+		fail.On(err != nil, "%s", err)
+
+		path := tree.WarrantyVoidedDir([]byte(common.ControllerIdentity()), []byte(common.HolotreeSpace))
+		return path, common.NewScorecard(), err
+	}
+
 	journal.CurrentBuildEvent().StartNow(force)
 
 	if settings.Global.NoBuild() {
@@ -68,6 +76,7 @@ func NewEnvironment(condafile, holozip string, restore, force bool, puller Catal
 
 	_, holotreeBlueprint, err := ComposeFinalBlueprint([]string{condafile}, "")
 	fail.On(err != nil, "%s", err)
+
 	common.EnvironmentHash, common.FreshlyBuildEnvironment = common.BlueprintHash(holotreeBlueprint), false
 	pretty.Progress(2, "Holotree blueprint is %q [%s with %d workers on %d CPUs].", common.EnvironmentHash, common.Platform(), anywork.Scale(), runtime.NumCPU())
 	journal.CurrentBuildEvent().Blueprint(common.EnvironmentHash)